# etc/docs/swagger.yaml
definitions:
...
  dto.CreateUserRequest:
    properties:
      age:
        maximum: 150
        minimum: 1
        type: integer
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        enum:
        - admin
        - user
        - guest
    required:
    - age
    - email
    - name
    - password
    - role
    type: object
  dto.HTTPErrorResp:
    properties:
      metadata:
        $ref: '#/definitions/dto.Meta'
    type: object
...
  dto.HttpSuccessResp:
    properties:
      data:
        x-order: "1"
      metadata:
        allOf:
        - $ref: '#/definitions/dto.Meta'
        x-order: "0"
      pagination:
        allOf:
        - $ref: '#/definitions/dto.Pagination'
        x-order: "2"
    type: object
...
  dto.Meta:
    properties:
      error:
        type: object
        x-order: "4"
      message:
        type: string
        x-order: "3"
      path:
        type: string
        x-order: "0"
      status:
        type: string
        x-order: "2"
      status_code:
        type: integer
        x-order: "1"
      timestamp:
        type: string
        x-order: "5"
    type: object
  dto.Pagination:
    properties:
      current_elements:
        type: integer
        x-order: "1"
      current_page:
        type: integer
        x-order: "0"
      cursor_end:
        type: string
        x-order: "7"
      cursor_start:
        type: string
        x-order: "6"
      sort_by:
        type: string
        x-order: "4"
      sort_dir:
        type: string
        x-order: "5"
      total_elements:
        type: integer
        x-order: "3"
      total_pages:
        type: integer
        x-order: "2"
    type: object
...
  dto.UpdateUserRequest:
    properties:
      age:
        maximum: 150
        minimum: 1
        type: integer
      email:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        enum:
        - admin
        - user
        - guest
    type: object
...
  entity.Role:
    enum:
    - admin
    - user
    - guest
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleGuest
  entity.User:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      updated_at:
        type: string
    type: object
host: localhost:8181
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Clean Architecture CRUD API with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go-Far
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get a paginated list of users with optional filters
      parameters:
      - description: Request cache control
        enum:
        - must-revalidate
        - must-db-revalidate
        in: header
        name: Cache-Control
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Minimum age
        in: query
        name: min_age
        type: integer
      - description: Maximum age
        in: query
        name: max_age
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction (asc/desc)
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HttpSuccessResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Delete user
      tags:
      - users
    get:
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.HttpSuccessResp'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPErrorResp'
      summary: Update user
      tags:
      - users
...
schemes:
- http
- https
swagger: "2.0"